Use UBI minimal instead of micro so we have proper OpenSSL packages
The ubi micro image doesn't have openssl packages and if we're using
that we're failing the FIPS check since the resulting container image
doesn't have the necessary openssl packages.

Let's use the minimal image which does that the necessary openssl
packages and should pass the resulting check-payload check to release
the operator.